update 5/1 16h

+++ src/main/java/vn/iotstar/controller/CategoryController.java
@@ -1,5 +1,6 @@
 package vn.iotstar.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import vn.iotstar.entity.Category;
+import vn.iotstar.entity.User;
 import vn.iotstar.service.CategoryService;
 
 @RestController
@@ -37,19 +39,39 @@ public class CategoryController {
 	@PostMapping("add")
 	public Category add(@RequestBody Category category) {
 		Category entity = category;
+		List<Category> list = cate.findAll();
+		
+		System.out.print("Name in "+ category.getName());
+		for (Category category2 : list) {
+			System.out.print("Name "+ category2.getName());
+			if(category2.getName().toString().equals(category.getName())) {
+				return null;
+			}
+		}
 		entity.setId(UUID.randomUUID().toString().split("-")[0]);
 		return cate.save(entity);
 	}
 	@PostMapping("update")
 	public Category update(@RequestBody Category category) {
-		
-		Category entity = category;
-		return cate.save(entity);
+		return cate.save(category);
 	}
 
 	@PostMapping("delete")
 	public String delete(@RequestParam(name = "id", required = false) String id) {
-		 cate.deleteById(id);
-		 return "Success";
+		List<Category> list = cate.findAll();
+		Boolean flag= false;
+		for (Category category : list) {
+			if(category.getId().equals(id)) {
+				flag = true;
+				break;
+			}
+		}
+		if(flag) {
+			cate.deleteById(id);
+			 return "Success";
+		}
+		else {
+			return null;
+		}
 	}
 }
